Support JWT for OnlyOffice
This commit adds a new configuration to send a JWT to OnlyOffice. It can
either be configured in the UI or set via the environment variable
`ONLYOFFICE_JWT_SECRET`.

To implement JWT handling a new dependency to
github.com/dgrijalva/jwt-go was introduced.

Even though the JWT check box will not be hidden when only office
integration is disabled the target was configured anyways. The token is
only applicable if the integration is enabled.

The `Valid` function for `OnlyOfficeConfigObject` had to be introduced
to satisfy the API of `jwt.NewWithClaims`.