Use Git repo to determine latest upstream commit
For some unclear reason the latest Docker image on DockerHub has the tag
`e72c0b4` but the latest commit on the master branch of the upstream
repo has the short hash `05fad99` (the previous commit `06f00bd` matches
the tag on DockerHub).

Use the latest commit from the repo now to determine what the base for
the auto-updated fork should be.